jobs/requirements.txt
fastapi
uvicorn
google-cloud-logging
google-cloud-bigquery
pydantic
python-dotenv
mixpanel
pinecone
openai
